libc: arm64: Use SIMD-accelerated memcpy
It turns out bionic is not using SIMD variant of memcpy although NEON can
be used in userspace. Import memcpy-advsimd from optimized-routines and
use it. Guard was added for extra safety but it is probably superfluous
on every modern Android device.

Initial commit in arm repo stated performance increase of 11%.

Signed-off-by: Andrzej Perczak <linux@andrzejperczak.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>